LORA发包时阻塞线程，阻止MCU进入休眠
注：MCU休眠时发送LORA，突然的大电流会引发死机

+++ src/devdrivers/sx1278/sx1278_hal.c
@@ -204,7 +204,7 @@ static void Board_sx1278LoraStatusCallback(PIN_Handle hPin, PIN_Id pinId)
  *
  * @return Pin status 
  */
-uint8_t Read_sx1278Dio0_Pin(void)
+int Read_sx1278Dio0_Pin(void)
 {
     return PIN_getInputValue(Board_SX1278_DIO0);
 }
